[Feat] 로그인 세션 활용 + DTO 수정

+++ src/main/java/kr/pe/fourj/controller/ReviewController.java
@@ -3,6 +3,8 @@ package kr.pe.fourj.controller;
 import java.time.LocalDateTime;
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -20,7 +22,6 @@ import kr.pe.fourj.exception.Exception.ArgumentNullException;
 import kr.pe.fourj.exception.Exception.NotFoundException;
 import kr.pe.fourj.service.CourseService;
 import kr.pe.fourj.service.ReviewService;
-import kr.pe.fourj.service.StudentService;
 
 @RestController
 public class ReviewController {
@@ -28,30 +29,32 @@ public class ReviewController {
 	@Autowired
 	private ReviewService reviewService;
 	@Autowired
-	private StudentService studentService;
-	@Autowired
 	private CourseService courseService;
 	
 	//후기 저장
 	@PostMapping("/review")
-	public ResponseDTO.Create saveReview(@RequestBody ReviewDTO.Create dto) {
+	public ResponseDTO.Create saveReview(HttpServletRequest request, @RequestBody ReviewDTO.Create dto) {
 		System.out.println("-- 후기 저장 시도 --");
 		
 		boolean result = false;
 		Long saveId = null;
-		try {
-			Student student = studentService.findOne(dto.getStudentIdx());
-			Course course = courseService.findOne(dto.getCourseIdx());
-			LocalDateTime dateTime = LocalDateTime.now();
-
+		if(request.getSession().getAttribute("student") != null) {
+			Object object = request.getSession().getAttribute("student");
+			Student entity = (Student)object;
+			
 			try {
-				saveId = reviewService.saveReview(new Review(student, course, dto.getContent(), dateTime, dto.getStar()));
-				result = true;
-			} catch (ArgumentNullException e) {
+				Course course = courseService.findOne(dto.getCourseIdx());
+				LocalDateTime dateTime = LocalDateTime.now();
+
+				try {
+					saveId = reviewService.saveReview(new Review(entity, course, dto.getContent(), dateTime, dto.getStar()));
+					result = true;
+				} catch (ArgumentNullException e) {
+					e.printStackTrace();
+				}
+			} catch (NotFoundException e) {
 				e.printStackTrace();
 			}
-		} catch (NotFoundException e) {
-			e.printStackTrace();
 		}
 
 		return new ResponseDTO.Create(saveId, result);
@@ -59,15 +62,27 @@ public class ReviewController {
 	
 	//후기 수정
 	@PutMapping("/review")
-	public ResponseDTO.Update updateReview(@RequestBody ReviewDTO.Update dto) {
+	public ResponseDTO.Update updateReview(HttpServletRequest request, @RequestBody ReviewDTO.Update dto) {
 		System.out.println("-- 후기 수정 시도 --");
 		
 		boolean result = false;
-		try {
-			reviewService.updateReview(dto);
-			result = true;
-		} catch (NotFoundException e) {
-			e.printStackTrace();
+		if(request.getSession().getAttribute("student") != null) {
+			Object object = request.getSession().getAttribute("student");
+			Student entity = (Student)object;
+			
+			List<Review> reviewList = entity.getReviewList();
+			
+			for(Review r : reviewList) {
+				if(r.getIdx() == dto.getIdx()) {
+					try {
+						reviewService.updateReview(dto);
+						result = true;
+					} catch (NotFoundException e) {
+						e.printStackTrace();
+					}
+					break;
+				}
+			}
 		}
 		
 		return new ResponseDTO.Update(result);
@@ -75,15 +90,27 @@ public class ReviewController {
 	
 	//후기 삭제
 	@DeleteMapping("/review")
-	public ResponseDTO.Delete deleteReview(@RequestBody ReviewDTO.Delete dto) {
+	public ResponseDTO.Delete deleteReview(HttpServletRequest request, @RequestBody ReviewDTO.Delete dto) {
 		System.out.println("-- 후기 삭제 시도 --");
 		
 		boolean result = false;
-		try {
-			reviewService.deleteReview(dto);
-			result = true;
-		} catch (NotFoundException e) {
-			e.printStackTrace();
+		if(request.getSession().getAttribute("student") != null) {
+			Object object = request.getSession().getAttribute("student");
+			Student entity = (Student)object;
+			
+			List<Review> reviewList = entity.getReviewList();
+			
+			for(Review r : reviewList) {
+				if(r.getIdx() == dto.getIdx()) {
+					try {
+						reviewService.deleteReview(dto);
+						result = true;
+					} catch (NotFoundException e) {
+						e.printStackTrace();
+					}
+					break;
+				}
+			}
 		}
 		
 		return new ResponseDTO.Delete(result);
@@ -116,25 +143,24 @@ public class ReviewController {
 		return new ResponseDTO.ReviewListResponse(true, reviewList);
 	}
 	
-	//특정 학생이 작성한 후기 리스트 검색
+	//특정 학생(자신)이 작성한 후기 리스트 검색
 	@GetMapping("/review/studentidx")
-	public ResponseDTO.ReviewListResponse findAllByStudentIdx(@RequestBody ReviewDTO.Get dto) {
+	public ResponseDTO.ReviewListResponse findAllByStudentIdx(HttpServletRequest request, @RequestBody ReviewDTO.Get dto) {
 		System.out.println("-- 특정 학생이 작성한 후기 리스트 검색 시도 --");
 		
 		boolean result = false;
 		List<Review> reviewList = null;
-		try {
-			Student student = studentService.findOne(dto.getStudentIdx());
-			reviewList = student.getReviewList();
+		if(request.getSession().getAttribute("student") != null) {
+			Object object = request.getSession().getAttribute("student");
+			Student entity = (Student)object;
+			reviewList = entity.getReviewList();
 			result = true;
-		} catch (NotFoundException e) {
-			e.printStackTrace();
 		}
 		
 		return new ResponseDTO.ReviewListResponse(result, reviewList);
 	}
 	
-	//특정 강의 이름으로 후기 리스트 검색
+	//특정 강의로 후기 리스트 검색
 	@GetMapping("/review/courseidx")
 	public ResponseDTO.ReviewListResponse findAllByCourseIdx(@RequestBody ReviewDTO.Get dto) {
 		System.out.println("-- 특정 학생이 작성한 후기 리스트 검색 시도 --");
